Fix world skills tab dependencies
