+++ client/src/components/WorldSkillsTab.jsx
@@ -5,9 +5,12 @@ import {
     ABILITY_DEFS,
     ABILITY_KEY_SET,
     NEW_WORLD_SKILL_ID,
+    SAVE_DEFS,
     WORLD_SKILL_SORT_OPTIONS,
     WORLD_SKILL_SORTERS,
+    abilityModifier,
     clampNonNegative,
+    formatModifier,
     makeCustomSkillId,
     normalizeCustomSkills,
     normalizeWorldSkillDefs,
@@ -16,6 +19,7 @@ import {
 } from "../constants/gameData";
 import { WORLD_SKILL_REFERENCE } from "../constants/referenceContent";
 import { get } from "../utils/object";
+import { deepClone, normalizeCharacter, normalizeSkills } from "../utils/character";
 
 import MathField from "./MathField";
